requirements.txt
slackclient==2.5.0
matplotlib>=3.1.3
Flask==1.1.1
